requirements/_base.txt
Django==1.5
South==0.7.6
Unipath==0.2.1
django-model-utils==1.2.0
psycopg2==2.4.6
wsgiref==0.1.2
